ban kernel codes from using the red zone
The red zone is an area specified by the AMD64 ABI, which is 128-byte
area beyond the location pointed to by RSP.  And the red zone must be
preserved over asynchronous events such as external interrupt and page
fault.

However, in kernel mode, the red zone is destroyed when an exception is
accepted because the stack is not switched and is modified for cpu to
save the current state.

Therefore, it is necessary to avoid use of the red zone in kernel mode.

refs #254