ihkmond: Fix location of kmsg buffer file
kmsg should be stored in a fast buffer before sent to /dev/log via
syslog(2) because syslog(2) drops messages when the message rate
exceeds the relatively low limit.
File is used as the buffer and its location is fixed so that it
resides in the node local file-system (i.e. /tmp/ihkmond).