ihk-smp: lookup unexported symbols at runtime
The only symbols still looked up at configure time are the arm64
gic_data and gic_raise_softirq symbols, because there are two with
the same name.
gic_data could be worked around and found with kallsyms_on_each_symbol()
because the gic_data_v2 structure is full of zeroes if not used,
but gic_raise_softirq is a function and both cannot be told appart
easily....

This means kaslr is not supported on arm right now, a future patch will
need to compute the offset and add that to the symbols from System.map
at run-time, or a smarter way needs to be found.

Change-Id: Ib6b4c8f24ef4a33b1d9e6690e2a338cedaa3a559