Update patch set 2
Patch Set 2:

I'm not sure I agree with this patch: if it's a real failure trying again won't help; if it's a trasient failure we need to understand why it failed and fix that instead. There is no reason ihkmond would be the only place that would run into this problem?

Patch-set: 2
CC: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>