Update patch set 1
Patch Set 1:

I think Iif ihk_ikc_packet is supposed to be the "superclass" of all IKC packets then it should be required as the first field and everywhere where ->channel is referenced (e.g., in the McKernel repo) the packet should be casted to (struct ihk_ikc_packet *) first. Otherwise it's not really a superclass.

Patch-set: 1
CC: Balazs Gerofi <1015205@d5d70762-12d0-45a1-890d-524b12d3f735>