smp request_irq: handle refcount better
Linux since 3.1.0 (b6873807a714 ("irq: Track the owner of irq
descriptor")) will take a module reference on request_irq and put it on
free_irq, but we want to free_irq on module exit, which cannot be called
if there are extra refs.

Additionally, since newer version of linux, the refcount in module exit
is -1 and try_module_get there will not do anything, making the free_irq
warn needlessly that it cannot put the module.

Instead of letting the irq manage the owner module, immediately put the
ref and unset the irq owner immediately after obtaining it - module_put
is smart enough to not do anything with a NULL module.

Change-Id: Ic9183e1220b93c058e312ad4dade014e472841ae