Update patch set 1
Patch Set 1:

> > high level comment: why is any of this arch-specific? All these remap functions should behave the same on x86, am I wrong?
> > If possible I'd rather move the x86 code to somewhere shared than duplicate code.
> 
> Since the x86 implementation was arch dependent, arm64 also implemented it.

Yes, but if the code itself is not "arch-dependent" (as in, only use base linux primitives that are generic) then it's much better to move the code out of x86 into the shared code than to copy it in the long run.
I did not compare the code in details but I think it should be easy enough to have a single implementation, please say if I missed something that makes it required to keep it in the arch-specific code.

Patch-set: 1
Reviewer: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>