Update patch set 3
Patch Set 3: Verified+1 Code-Review+2

> It does not occur in this environment, so I'm not sure.

I can confirm this boots on an arm VM, so you'd need access to real hardware to debug this... sigh.

Well, it turns out it was simpler than expected thanks to having split the get_processor_id() change - that was the problem. For some reason on apollo that fails so the first mckernel core doesn't match with itself and it tries to boot itself, which does not work.
If I only take the first patch then it works, so happy with this now - I'll block the other patch instead.

Patch-set: 3
Reviewer: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2
Label: Verified=+1