ihklib: ihk_reserve_mem: Reserve by specifying total amount
Try to reserve the sum of the requested per-NUMA-node amounts in a
balanced way. It reports an error when the variance of the reserved
amounts of the NUMA-nodes exceeds the specified limit.

The usage steps are:
(1) Call ihk_reserve_mem_conf(<device id>, IHK_RESERVE_MEM_TOTAL,
    <variance limit in percent>)
(2) Call ihk_reserve_mem()

Change-Id: I29256f317e6ff3a13c2ff22cb7abf26b76e534ec